moved some helper functions to utils
