+++ CaloCem/utils.py
@@ -6,7 +6,7 @@ from scipy.interpolate import InterpolatedUnivariateSpline, UnivariateSpline
 from scipy.signal import convolve, gaussian
 from scipy.ndimage import median_filter
 from pathlib import Path
-
+import re
 
 def create_base_plot(data, ax, _age_col, _target_col, sample):
     """
@@ -123,6 +123,76 @@ def fit_univariate_spline(df, target_col, s=1e-6):
     return df
 
 
+def remove_unnecessary_data(df):
+    # cut out data part
+    df = df.iloc[1:, :].reset_index(drop=True)
+
+    # drop column
+    try:
+        data = df.drop(columns=["time_markers_nan"])
+    except KeyError:
+        pass
+
+    # remove columns with too many NaNs
+    data = data.dropna(axis=1, thresh=3)
+
+    # # remove rows with NaNs
+    data = data.dropna(axis=1)
+
+    return data
+
+
+def add_sample_info(df, file):
+
+    # get sample name
+    sample_name = Path(file).stem
+
+    # add sample information
+    df["sample"] = file
+    # df["sample_short"] = sample_name
+    df.assign(sample_short=sample_name)
+
+    return df
+
+def tidy_colnames(df):
+    # get new column names
+    new_columnames = []
+    for i in df.iloc[0, :]:
+        # build
+        new_columname = (
+            re.sub(r'[\s\n\[\]\(\)° _"]+', "_", i.lower())
+            .replace("/", "_")
+            .replace("_signal_", "_")
+            .strip("_")
+        )
+
+        # select appropriate unit
+        if new_columname == "time":
+            new_columname += "_s"
+        elif "temperature" in new_columname:
+            new_columname += "_c"
+        elif new_columname == "heat_flow":
+            new_columname += "_w"
+        elif new_columname == "heat":
+            new_columname += "_j"
+        elif new_columname == "normalized_heat_flow":
+            new_columname += "_w_g"
+        elif new_columname == "normalized_heat":
+            new_columname += "_j_g"
+        else:
+            new_columname += "_nan"
+
+        # add to list
+        new_columnames.append(new_columname)
+
+    # set
+    df.columns = new_columnames
+    # validate new column names
+    if not "time_s" in new_columnames:
+        # stop here
+        return None
+
+    return df
 # def calculate_smoothed_heatflow_derivatives(
 #     df: pd.DataFrame,
 #     tianparams: TianParameters,
